Started creating mobile views and splash page
