+++ src/components/Browser/SplashPage.jsx
@@ -1,0 +1,7 @@
+const SplashPage = () => {
+  return (
+    <div>START HERE</div>
+  )
+}
+
+export default SplashPage;
