+++ src/components/Mobile/MobileSplashPage.jsx
@@ -1,0 +1,7 @@
+const MobileSplashPage = () => {
+  return (
+    <div>START HERE</div>
+  )
+}
+
+export default MobileSplashPage;
